Move auth to a new route & Refactor some names
